Support larger ints in fibonacci program

+++ programs/c/fibonacci.c
@@ -1,8 +1,9 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 
-int fib(int index) {
+uint64_t fib(int index) {
     if (index <= 1) return index;
     else return fib(index - 1) + fib(index - 2);
 }
@@ -12,6 +13,6 @@ int main(int argc, char *argv[]) {
         printf("Not enough arguments!\n");
         exit(1);
     }
-    unsigned long long index = atoll(argv[1]);
+    uint64_t index = atoll(argv[1]);
     printf("%llu\n", fib(index));
 }
